Update README.md
Use of less-than and greater-than symbols throughout README.md is inappropriate for Markdown; the content will be rendered as raw HTML.  This commit fixes all of this by using backticks around the content (which is also partially used throughout the same document).  I simply did a find/search for less-than and manually made the changes.